+++ sdk/nodejs/output.ts
@@ -69,8 +69,12 @@ class OutputImpl<T> implements OutputInstance<T> {
      * The list of resource that this output value depends on.
      *
      * Only callable on the outside.
+     *
+     * Normally returns a Promise<Set>.  However, is typed to return a Set as well to properly
+     * represent and force us to handle values produced by older versions of the sdk in SxS
+     * scenarios.
      */
-    public readonly resources: () => Set<Resource>;
+    public readonly resources: () => Set<Resource> | Promise<Set<Resource>>;
 
     /**
      * [toString] on an [Output<T>] is not supported.  This is because the value an [Output] points
@@ -137,31 +141,25 @@ class OutputImpl<T> implements OutputInstance<T> {
 
     /** @internal */
     public constructor(
-            resources: Set<Resource> | Resource[] | Resource,
+            resources: Set<Resource> | Resource[] | Resource | Promise<Set<Resource> | Resource[] | Resource>,
             promise: Promise<T>,
             isKnown: Promise<boolean>,
             isSecret: Promise<boolean>) {
+
+        // Always create a copy so that no one accidentally modifies our Resource list.
+        const resourcesSet = Promise.resolve(resources).then(
+            rs => Array.isArray(rs) ? new Set(rs) :
+                  rs instanceof Set ? new Set(rs) :
+                  new Set([rs]));
 
         // We are only known if we are not explicitly unknown and the resolved value of the output
         // contains no distinguished unknown values.
         this.isKnown = Promise.all([isKnown, promise]).then(([known, val]) => known && !containsUnknowns(val));
         this.isSecret = isSecret;
 
-        let resourcesArray: Resource[];
-
-        // Always create a copy so that no one accidentally modifies our Resource list.
-        if (Array.isArray(resources)) {
-            resourcesArray = resources;
-        } else if (resources instanceof Set) {
-            resourcesArray = [...resources];
-        } else {
-            resourcesArray = [resources];
-        }
-
-        this.resources = () => new Set<Resource>(resourcesArray);
+        this.resources = () => resourcesSet;
         this.promise = (withUnknowns?: boolean) => OutputImpl.getPromisedValue(promise, withUnknowns);
 
-        const firstResource = resourcesArray[0];
         this.toString = () => {
             const message =
 `Calling [toString] on an [Output<T>] is not supported.
@@ -269,17 +267,20 @@ To manipulate the value of this Output, use '.apply' instead.`);
     // advantage of this to allow proxied property accesses to return known values even if other properties of
     // the containing object are unknown.
     public apply<U>(func: (t: T) => Input<U>, runWithUnknowns?: boolean): Output<U> {
-        const applied = Promise.all([this.promise(/*withUnknowns*/ true), this.isKnown, this.isSecret])
-                               .then(([value, isKnown, isSecret]) => applyHelperAsync<T, U>(value, isKnown, isSecret, func, !!runWithUnknowns));
+        const applied = Promise.all([this.resources(), this.promise(/*withUnknowns*/ true), this.isKnown, this.isSecret])
+                               .then(([resources, value, isKnown, isSecret]) => applyHelperAsync<T, U>(resources, value, isKnown, isSecret, func, !!runWithUnknowns));
 
         const result = new OutputImpl<U>(
-            this.resources(), applied.then(a => a.value), applied.then(a => a.isKnown), applied.then(a => a.isSecret));
+            applied.then(a => a.resources), applied.then(a => a.value),
+            applied.then(a => a.isKnown), applied.then(a => a.isSecret));
         return <Output<U>><any>result;
     }
 }
 
 // tslint:disable:max-line-length
-async function applyHelperAsync<T, U>(value: T, isKnown: boolean, isSecret: boolean, func: (t: T) => Input<U>, runWithUnknowns: boolean) {
+async function applyHelperAsync<T, U>(
+        resources: Set<Resource>, value: T, isKnown: boolean, isSecret: boolean,
+        func: (t: T) => Input<U>, runWithUnknowns: boolean) {
     if (runtime.isDryRun()) {
         // During previews only perform the apply if the engine was able to give us an actual value
         // for this Output.
@@ -288,6 +289,7 @@ async function applyHelperAsync<T, U>(value: T, isKnown: boolean, isSecret: bool
         if (!applyDuringPreview) {
             // We didn't actually run the function, our new Output is definitely **not** known.
             return {
+                resources,
                 value: <U><any>undefined,
                 isKnown: false,
                 isSecret,
@@ -305,18 +307,23 @@ async function applyHelperAsync<T, U>(value: T, isKnown: boolean, isSecret: bool
     const transformed = await func(value);
     if (Output.isInstance(transformed)) {
         // Note: if the func returned a Output, we unwrap that to get the inner value returned by
-        // that Output.  Note that we are *not* capturing the Resources of this inner Output.
-        // That's intentional.  As the Output returned is only supposed to be related this *this*
-        // Output object, those resources should already be in our transitively reachable resource
-        // graph.
+        // that Output.  Note that we *are* capturing the Resources of this inner Output and lifting
+        // them up to the outer Output as well.
 
         // Note: we intentionally await all the promises of the transformed value.  This way we
-        // properly propogate any rejections of any of them through ourselves as well.
+        // properly propagate any rejections of any of them through ourselves as well.
         const innerValue = await transformed.promise(/*withUnknowns*/ true);
         const innerIsKnown = await transformed.isKnown;
         const innerIsSecret = await (transformed.isSecret || Promise.resolve(false));
+        const innerResources = await transformed.resources();
+
+        const totalResources = new Set(resources);
+        for (const res of innerResources) {
+            totalResources.add(res);
+        }
 
         return {
+            resources: totalResources,
             value: innerValue,
             isKnown: innerIsKnown,
             isSecret: isSecret || innerIsSecret,
@@ -326,6 +333,7 @@ async function applyHelperAsync<T, U>(value: T, isKnown: boolean, isSecret: bool
     // We successfully ran the inner function.  Our new Output should be considered known.  We
     // preserve secretness from our original Output to the new one we're creating.
     return {
+        resources,
         value: transformed,
         isKnown: true,
         isSecret,
@@ -454,7 +462,7 @@ export function all<T>(val: Input<T>[] | Record<string, Input<T>>): Output<any> 
         const [resources, isKnown, isSecret] = getResourcesAndDetails(allOutputs);
         const promisedArray = Promise.all(allOutputs.map(o => o.promise(/*withUnknowns*/ true)));
 
-        return new Output<Unwrap<T>[]>(new Set<Resource>(resources), promisedArray, isKnown, isSecret);
+        return new Output<Unwrap<T>[]>(resources, promisedArray, isKnown, isSecret);
     } else {
         const keysAndOutputs = Object.keys(val).map(key => ({ key, value: output(val[key]) }));
         const allOutputs = keysAndOutputs.map(kvp => kvp.value);
@@ -462,7 +470,7 @@ export function all<T>(val: Input<T>[] | Record<string, Input<T>>): Output<any> 
         const [resources, isKnown, isSecret] = getResourcesAndDetails(allOutputs);
         const promisedObject = getPromisedObject(keysAndOutputs);
 
-        return new Output<Record<string, Unwrap<T>>>(new Set<Resource>(resources), promisedObject, isKnown, isSecret);
+        return new Output<Record<string, Unwrap<T>>>(resources, promisedObject, isKnown, isSecret);
     }
 }
 
@@ -476,16 +484,26 @@ async function getPromisedObject<T>(
     return result;
 }
 
-function getResourcesAndDetails<T>(allOutputs: Output<Unwrap<T>>[]): [Resource[], Promise<boolean>, Promise<boolean>] {
-    const allResources = allOutputs.reduce<Resource[]>((arr, o) => (arr.push(...o.resources()), arr), []);
+function getResourcesAndDetails<T>(allOutputs: Output<Unwrap<T>>[]): [Promise<Set<Resource>>, Promise<boolean>, Promise<boolean>] {
+    const resources = Promise.all(allOutputs.map(o => o.resources())).then(arr => {
+        const resourceSet = new Set<Resource>();
+
+        for (const set of arr) {
+            for (const res of set) {
+                resourceSet.add(res);
+            }
+        }
+
+        return resourceSet;
+    });
 
     // A merged output is known if all of its inputs are known.
     const isKnown = Promise.all(allOutputs.map(o => o.isKnown)).then(ps => ps.every(b => b));
 
     // A merged output is secret if any of its inputs are secret.
-    const isSecret = Promise.all(allOutputs.map(o => isSecretOutput(o))).then(ps => ps.find(b => b) !== undefined);
+    const isSecret = Promise.all(allOutputs.map(o => isSecretOutput(o))).then(ps => ps.some(b => b));
 
-    return [allResources, isKnown, isSecret];
+    return [resources, isKnown, isSecret];
 }
 
 /**
@@ -634,7 +652,7 @@ export interface OutputInstance<T> {
     /** @internal */ readonly isKnown: Promise<boolean>;
     /** @internal */ readonly isSecret: Promise<boolean>;
     /** @internal */ promise(withUnknowns?: boolean): Promise<T>;
-    /** @internal */ resources(): Set<Resource>;
+    /** @internal */ resources(): Set<Resource> | Promise<Set<Resource>>;
 
     /**
      * Transforms the data of the output with the provided func.  The result remains a
@@ -651,7 +669,7 @@ export interface OutputInstance<T> {
      * ```
      *
      * In this example, taking a dependency on d2 means a resource will depend on all the resources
-     * of d1.  It will *not* depend on the resources of v.x.y.OtherDep.
+     * of d1.  It will *also* depend on the resources of v.x.y.OtherDep.
      *
      * Importantly, the Resources that d2 feels like it will depend on are the same resources as d1.
      * If you need have multiple Outputs and a single Output is needed that combines both
@@ -689,7 +707,7 @@ export interface OutputConstructor {
     isInstance<T>(obj: any): obj is Output<T>;
 
     /** @internal */ new<T>(
-            resources: Set<Resource> | Resource[] | Resource,
+            resources: Set<Resource> | Resource[] | Resource | Promise<Set<Resource> | Resource[] | Resource>,
             promise: Promise<T>,
             isKnown: Promise<boolean>,
             isSecret: Promise<boolean>): Output<T>;